👌🚚 [Core] rename core errorCollection
