+++ packages/core/src/domain/internalMonitoring.ts
@@ -2,8 +2,8 @@
 import { combine, Context } from '../tools/context'
 import * as utils from '../tools/utils'
 import { Batch, HttpRequest } from '../transport/transport'
+import { toStackTraceString } from './automaticErrorCollection'
 import { Configuration } from './configuration'
-import { toStackTraceString } from './errorCollection'
 import { computeStackTrace } from './tracekit'
 
 enum StatusType {
